added support for multiple serialization formats
